feat(testworkflows): move Cloud TestWorkflow instructions outside of TCL
