+++ cmd/testworkflow-toolkit/artifacts/handler_test.go
@@ -6,9 +6,9 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/kubeshop/testkube/pkg/tcl/cloudtcl/data/testworkflow"
-
 	"github.com/golang/mock/gomock"
+
+	"github.com/kubeshop/testkube/pkg/cloud/data/testworkflow"
 
 	"github.com/kubeshop/testkube/cmd/testworkflow-toolkit/common/testdata"
 	"github.com/kubeshop/testkube/pkg/cloud/data/artifact"
