+++ cmd/testworkflow-toolkit/commands/artifacts_test.go
@@ -8,9 +8,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/kubeshop/testkube/pkg/cloud/data/testworkflow"
 	"github.com/kubeshop/testkube/pkg/utils/test"
-
-	"github.com/kubeshop/testkube/pkg/tcl/cloudtcl/data/testworkflow"
 
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
